Move the blockLogView into the block_viewer.go
